🚚(sso) move silentLogin code with agentConnect related code
makes more sense here than having it in the "utils" file
