+++ lib/web/baseRouter.js
@@ -3,7 +3,7 @@
 const Router = require('express').Router
 
 const response = require('../response')
-const { silentLogin } = require('./utils')
+const silentLogin = require('./auth/agentConnect/silentLogin')
 
 const baseRouter = module.exports = Router()
