+++ lib/web/note/router.js
@@ -1,7 +1,8 @@
 'use strict'
 
 const Router = require('express').Router
-const { markdownParser, silentLogin } = require('../utils')
+const { markdownParser } = require('../utils')
+const silentLogin = require('../auth/agentConnect/silentLogin')
 
 const router = module.exports = Router()
